Moved test classes from test into fixtures directory
