+++ test/bootstrap.php
@@ -6,6 +6,6 @@ call_user_func(function() {
     }
 
     $loader = require $autoloadFile;
-    $loader->add('BDBStudios\BreakfastSerializer\Tests', __DIR__);
-
+    $loader->add('BDBStudios\BreakfastSerializerTests', __DIR__);
+    $loader->add('BDBStudios\BreakfastSerializerTests\Fixtures', __DIR__ . '/Fixtures');
 });
